Added tracing points to QUndoStack so that it can be undone/redone
Created the TracingPointsAddCommand that keeps track of points added to
the stack. It is a very efficient method of keeping track of the points
since there are likely to be many of these commands in one use of this
application.

+++ coronalslicewidget.cpp
@@ -405,7 +405,7 @@ void CoronalSliceWidget::mousePressEvent(QMouseEvent *eventPress)
         // Flag to indicate that panning is occuring
         // The starting position is stored so to know how much movement has occurred
         startPan = true;
-        moveID = ((moveID + 1) > CommandID::CoronalMoveEnd) ? CommandID::CoronalMove : (CommandID)(moveID + 1);
+        moveID = (((int)moveID + 1) > (int)CommandID::CoronalMoveEnd) ? CommandID::CoronalMove : (CommandID)((int)moveID + 1);
         lastMousePos = eventPress->pos();
     }
 }
